merge release v0.3.2 into dev

+++ ELA/core/types/payload/dposillegalproposals.go
@@ -46,7 +46,7 @@ func (d *ProposalEvidence) Deserialize(r io.Reader) (err error) {
 		return err
 	}
 
-	if d.BlockHeader, err = common.ReadVarBytes(r, uint32(pact.MaxBlockSize),
+	if d.BlockHeader, err = common.ReadVarBytes(r, pact.MaxBlockSize,
 		"block header"); err != nil {
 		return err
 	}
